rename allocator.rs to allocator_impl.rs
